requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
jinja2
torch
torchvision
pillow